+++ src/containers/Home/actions/index.js
@@ -1,0 +1,8 @@
+import fetchMyClashesAction from "./fetchMyClashesAction";
+import fetchRecentClashesAction from "./fetchRecentClashesAction";
+
+export {
+  fetchMyClashesAction,
+  fetchRecentClashesAction
+}
+
